ContextUp_sandbox/requirements.txt
# ContextUp Consolidated Requirements
# Python Version: 3.11.x (FULL installer required, NOT embeddable - tkinter needed)
# Usage: pip install -r requirements.txt

# --- Core UI & System ---
customtkinter
psutil
pystray
packaging
pywin32
requests
pynput
keyboard
send2trash
piexif
pyperclip  # Clipboard handling
xxhash  # Fast hashing for Finder (10x faster than MD5)
PySide6
moderngl
glcontext

# --- Vectorizer (RigReady) ---
vtracer
psd-tools

# --- Multimedia & File Handling ---
Pillow
ffmpeg-python
yt-dlp
pypdf
pdf2image
imageio  # EXR/image sequence handling
OpenEXR  # EXR file format support
opencv-python
scipy  # Image/signal processing
svglib  # SVG to Image conversion
reportlab  # SVG rendering dependency
pdf2docx  # PDF to Word
pymupdf  # PDF processing (images, text, tables) - no Poppler needed
pymupdf4llm  # PDF to Markdown (LLM optimized)
openpyxl  # PDF to Excel
python-pptx  # PDF to PowerPoint
comtypes  # Office to PDF (Windows only)
markdown  # Markdown processing
xhtml2pdf  # HTML/Markdown to PDF
ebooklib  # PDF to EPUB

# --- AI & Processing ---
setuptools
wheel
packaging
numpy<2
noise
tqdm
rembg
numba  # JIT acceleration for numeric kernels
cupy-cuda12x  # GPU backend for Noise Master (CUDA 12.x)
accelerate  # Hugging Face model loading
diffusers  # Stable Diffusion pipelines
transformers # Required for diffusion pipelines
huggingface_hub

# --- PyTorch (CUDA 12.1) ---
# Pin versions for compatibility with gfpgan/basicsr (removed functional_tensor in 0.17)
# torch 2.1.2 + torchvision 0.16.2 is the last stable combination that works with basicsr
torch==2.1.2 --index-url https://download.pytorch.org/whl/cu121
torchvision==0.16.2 --index-url https://download.pytorch.org/whl/cu121
torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu121

# --- AI Models & APIs ---
open_clip_torch
faster-whisper
google-generativeai
google-genai  # Google GenAI SDK v2 (Gemini 1.5+)
deep-translator
kornia  # Computer Vision (Required for RMBG)

# --- 3D Processing ---
pymeshlab  # Mesh LOD generation
rawpy  # RAW image processing (Canon CR2, Nikon NEF, Sony ARW, etc.)
pillow-heif  # HEIC/AVIF support

# --- Audio Processing ---
pygame  # Required for ACE Audio Editor playback
demucs  # Audio stem separation (vocals, drums, bass, other)

# --- Text Processing ---


# --- AI Models & APIs ---
ollama  # Required for local LLM/Vision tasks
gfpgan  # Face restoration for Upscale

# --- Optional / Experimental ---
basicsr
realesrgan
